requirements.txt
absl-py==1.0.0
asgiref==3.4.1
black==21.10b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Django==3.2.9
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.41.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
pandas==1.3.4
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.2
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
